removing bad project and renamed gatereader class
