+++ src/Devices/GateReader/Transportation.Demo.Devices.GateReader.csproj
