--- src/Devices/Transportation.Demo.Devices.GateReader/Sample.AppSettings.json
@@ -1,6 +1,0 @@
-{
-    "AppSettings": 
-    {
-      "IoTConnectionString": "PutHereYourConnectionString"
-    }
-}
